Request all details before code sending
